refactor: create reusable hook to intercept notifications
